pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["CM360 <31625347+cm-360@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
quart = "^0.19.9"
ollama = "^0.4.4"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app:run"

[tool.pyright]
venvPath = "."
venv = ".venv"
